squish files together into buildable version
