--- cmd/run-webserver/local.go
@@ -1,13 +1,0 @@
-//+build local
-
-package main
-
-func SetupFPGA() {
-}
-
-func GetMD5HashFPGA(text string) string {
-	return GetMD5Hash(text)
-}
-
-func CleanupFPGA() {
-}
